Create the class files
